Creation du transaction-history-microservice

+++ backend/product-service/src/controller/product.controller.js
@@ -1,4 +1,6 @@
 const Product = require("../models/Product");
+const { TRANSACTION_HISTOY_API_URL } = require('../microserviceURL/microserviceUrl');
+const axios = require('axios');
 
 const getAll = async (req, res) => {
     try {
@@ -9,14 +11,23 @@ const getAll = async (req, res) => {
     }
   };
 
-const createProduct = async (req, res) => {
-    try {
-      const product = await Product.create(req.body);
-      res.status(201).json(product);
-    } catch (err) {
-      res.status(500).json({ message: "server error Product has not been created" });
-    }
+  const createProduct = async (req, res) => {
+      try {
+          const product = await Product.create(req.body);
+          const data = {
+              transactionType: "Achat",
+              productId: product.id,
+              productName: req.body.name,
+              price: req.body.price,
+          };
+          await axios.post(`${TRANSACTION_HISTOY_API_URL}/transaction-history/create`, data);
+          res.status(201).json(product);
+      } catch (err) {
+          console.error(err);
+          res.status(500).json({ message: "Server error: Product has not been created" });
+      }
   };
+  
 
 const updateProduct = async (req, res) => {
   try {
@@ -28,16 +39,28 @@ const updateProduct = async (req, res) => {
     }
   };
 
-const deleteProduct = async (req, res) => {
+  const deleteProduct = async (req, res) => {
     const id = req.params.id;
-
     try {
-      const product = await Product.destroy({ where: { id: id } });
-      res.status(200).json(product);
+        const product = await Product.findByPk(id);
+        if (!product) {
+            return res.status(404).json({ message: "Product not found" });
+        }
+        const data = {
+            transactionType: "Vente",
+            productId: product.id,
+            productName: product.name,
+            price: product.price,
+        };
+        await axios.post(`${TRANSACTION_HISTOY_API_URL}/transaction-history/create`, data);
+        await Product.destroy({ where: { id: id } });
+
+        res.status(200).json({ message: "Product deleted successfully" });
     } catch (err) {
-      res.status(500).json({ message: "Product not found" });
+        console.error(err);
+        res.status(500).json({ message: "Server error: Product could not be deleted" });
     }
-  };
+};
 
   module.exports = {
     getAll,
